+++ app/controllers/registrations_controller.rb
@@ -2,6 +2,7 @@
 
 class RegistrationsController < Devise::RegistrationsController
   skip_before_action :verify_authenticity_token, only: [:create]
+  prepend_before_action :check_captcha, only: [:create]
 
   protected
 
@@ -13,5 +14,12 @@ class RegistrationsController < Devise::RegistrationsController
 
   def sign_up_params
     params.require(:user).permit(:first_name, :last_name, :over_eighteen, :email, :password, :beta_opt_in, :invite_token, :provider_id)
+  end
+
+  def check_captcha
+    if RECAPTCHA_ENABLED && !verify_recaptcha
+      self.resource = resource_class.new sign_up_params
+      respond_with_navigational(resource) { render :new }
+    end
   end
 end