+++ config/routes.rb
@@ -302,7 +302,10 @@ Rails.application.routes.draw do
 
   get 'admin/social-media', to: 'admin#social_media'
 
-  devise_for :users, controllers: { registrations: 'registrations', sessions: 'sessions' }, path_names: { sign_up: 'join', sign_in: 'login' }, path: ''
+  devise_for :users,
+             controllers: { registrations: 'registrations', sessions: 'sessions' },
+             path_names: { sign_up: 'join', sign_in: 'login' },
+             path: ''
 
   resources :users do
     collection do